removed unnecessary toast (#2036)
