+++ extensions/nextjs-docs/CHANGELOG.md
@@ -1,9 +1,8 @@
-# Next.js Changelog
+# nextjs Changelog
 
-## [Example Search] - 2022-04-06
-- Add `Search Examples` command
-- Rename `Search Next.js Documentation` to `Search Documentation`
-- Add `Next.js` as subtitle for all commands
+## [Performance Improvement] - 2022-05-24
+
+- Performance improvement
+- Caching improvement
 
 ## [Initial Version] - 2022-04-06
-- Add `Search Next.js Documentation` command
