+++ extensions/nextjs-docs/package.json
@@ -5,7 +5,6 @@
   "description": "Efficiently search Next.js documentation",
   "icon": "command-icon.png",
   "author": "PrasadChinwal",
-  "contributors": ["vimtor"],
   "categories": [
     "Productivity",
     "Documentation",
@@ -14,17 +13,9 @@
   "license": "MIT",
   "commands": [
     {
-      "name": "search-documentation",
-      "title": "Search Documentation",
-      "subtitle": "Next.js",
+      "name": "index",
+      "title": "Search Next.js Documentation",
       "description": "Search the Next.js documentation",
-      "mode": "view"
-    },
-    {
-      "name": "search-examples",
-      "title": "Search Examples",
-      "subtitle": "Next.js",
-      "description": "Search examples provided by Next.js",
       "mode": "view"
     }
   ],
